debug trace, progmem, broken!

+++ snack_style_organics/support/pk_pid_ctl/pk_pid_ctl_code/_src/pid_ctl.cpp
@@ -123,7 +123,7 @@ void Pid_Ctl_Read_FeedBack(const uint8_t channel)
 bool Wait_For_Sample(void)
 {
    uint32_t time_now = pid::Get_Sample_Time();
-   TR_INFO_2("Waiting %d == %d", time_now, Sample_Tout);
+   TR_INFO_2("Waiting %d == %d", (long) time_now, (long) Sample_Tout);
    return ( (time_now -  Sample_Tout) >= PID_CTL_TAU_COEFF_MS);
 }
 /*=====================================================================================* 
@@ -148,7 +148,7 @@ void pid::Init(void)
 
 void pid::Set_Point(const PID_CHANNEL_T channel, const Fix32_T val)
 {
-   TR_INFO_3("%s pid_ch = %d, val %ld", __FUNCTION__,channel,val);
+   TR_INFO_3("%s pid_ch = %d, val %ld", __FUNCTION__,channel,(int32_t)val);
    Pid_Channels[channel].set_point = val;
 }
